backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
pytesseract==0.3.13
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
